Rewrite CPU to use multiple lookup tables
* s_cpu_bits stores all 64-bit micro-op values decoded in cpu_step.
* s_opcode_bits stores a sequence of micro-ops for each of the
  instructions, as well as the IRQ/NMI/DMA sequences.
* s_opcode_loc maps from an opcode to the index in s_opcode_bits

It seems like this might be slower, and it does seem to run more
instructions, but consistently has a higher insn-per-cycle count.

+++ src/emulator.h
@@ -95,7 +95,7 @@ typedef struct {
 
 typedef struct {
   u64 bits;
-  const u64 *step, *next_step;
+  const u8 *step, *next_step;
   u8 PCL, PCH, TL, TH, buslo, bushi, fixhi, veclo, oamhi;
   u8 A, X, Y, S;
   u8 ram[0x800], prg_ram[0x2000];
